Adding a base controller
