+++ SecretSanta/test/SecretSanta.Api.Tests/Controllers/BaseControllerTests.cs
@@ -1,0 +1,41 @@
+﻿using AutoMapper;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using SecretSanta.Business;
+using SecretSanta.Data;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Net.Http;
+using System.Threading.Tasks;
+
+namespace SecretSanta.Api.Tests.Controllers
+{
+    [TestClass]
+    public class BaseControllerTests
+    {
+        //Initialized in the test initialization
+#nullable disable
+        protected SecretSantaWebApplicationFactory Factory { get; set; }
+        protected HttpClient Client { get; set; }
+#nullable enable
+
+        protected IMapper Mapper { get; } = AutomapperConfigurationProfile.CreateMapper();
+
+        [TestInitialize]
+        public void TestSetup()
+        {
+            Factory = new SecretSantaWebApplicationFactory();
+
+            using ApplicationDbContext context = Factory.GetDbContext();
+            context.Database.EnsureCreated();
+
+            Client = Factory.CreateClient();
+        }
+
+        [TestCleanup]
+        public void TestCleanup()
+        {
+            Factory.Dispose();
+        }
+    }
+}
